Rename file to wrapper
