+++ lib/wrapper.js
@@ -103,18 +103,17 @@ class Tester {
     });
   }
 
-  dumpTrace(callback) {
+  dumpTrace() {
     if (noflo.isBrowser() || !this.tracer) {
-      callback(null);
-      return;
+      return Promise.resolve();
     }
-    writeTrace({
+    return writeTrace({
       baseDir: this.baseDir,
     }, this.tracer)
       .then((f) => {
         console.log('Wrote flowtrace to', f);
-        callback(null);
-      }, callback);
+        return f;
+      });
   }
 
   // Sends data packets to one or multiple inports and disconnects them.
