+++ package.json
@@ -2,7 +2,7 @@
   "name": "noflo-wrapper",
   "version": "0.3.0",
   "description": "NoFlo component/graph testing and embedding wrapper compatible with BDD/TDD/whatever",
-  "main": "lib/tester.js",
+  "main": "lib/wrapper.js",
   "scripts": {
     "pretest": "eslint lib/*.js spec/*.js",
     "test": "./node_modules/.bin/mocha spec/*.js"
